Changed voxel type from signed to unsigned.
Having voxels be 'char' was leading to some strange bugs with the voxel
indices. I'm surprised there were more than 127 different types of
voxels in one scene (it probably went over once it started storing all
the chasm permutations). Extending the limit to 255 should be fine,
since extending it to 65535 (two bytes per voxel) would otherwise be a
waste of memory.

+++ OpenTESArena/src/World/LevelData.h
@@ -53,7 +53,7 @@ private:
 	VoxelGrid voxelGrid;
 	double ceilingHeight;
 
-	void setVoxel(int x, int y, int z, int id);
+	void setVoxel(int x, int y, int z, uint8_t id);
 	void readFLOR(const std::vector<uint8_t> &flor, int width, int depth,
 		const INFFile &inf);
 	void readMAP1(const std::vector<uint8_t> &map1, int width, int depth,
